Migrate internals not related to flux to React.Internal module
